메인 화면의 각 Fragment 비중이 높으므로 Package로 그룹화 완료
