+++ solution/app/src/main/java/com/smparkworld/docfriends/ui/main/MainActivity.kt
@@ -6,6 +6,11 @@ import androidx.databinding.DataBindingUtil
 import com.smparkworld.docfriends.R
 import com.smparkworld.docfriends.databinding.ActivityMainBinding
 import com.smparkworld.docfriends.extension.setupWithNavController
+import com.smparkworld.docfriends.ui.main.care.CareFragment
+import com.smparkworld.docfriends.ui.main.chat.ChatFragment
+import com.smparkworld.docfriends.ui.main.home.HomeFragment
+import com.smparkworld.docfriends.ui.main.me.MeFragment
+import com.smparkworld.docfriends.ui.main.video.VideoFragment
 
 class MainActivity : AppCompatActivity() {
