+++ solution/app/src/main/java/com/smparkworld/docfriends/ui/main/home/HomeFragment.kt
@@ -1,9 +1,8 @@
-package com.smparkworld.docfriends.ui.main
+package com.smparkworld.docfriends.ui.main.home
 
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.smparkworld.docfriends.databinding.FragmentMainHomeBinding
