+++ solution/app/src/main/java/com/smparkworld/docfriends/ui/main/me/MeFragment.kt
@@ -1,4 +1,4 @@
-package com.smparkworld.docfriends.ui.main
+package com.smparkworld.docfriends.ui.main.me
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,11 +6,11 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.smparkworld.docfriends.R
 
-class ChatFragment : Fragment() {
+class MeFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = inflater.inflate(R.layout.fragment_main_chat, container, false)
+    ) = inflater.inflate(R.layout.fragment_main_me, container, false)
 }
