+++ solution/app/src/main/java/com/smparkworld/docfriends/ui/main/video/VideoFragment.kt
@@ -1,4 +1,4 @@
-package com.smparkworld.docfriends.ui.main
+package com.smparkworld.docfriends.ui.main.video
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,11 +6,11 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.smparkworld.docfriends.R
 
-class MeFragment : Fragment() {
+class VideoFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = inflater.inflate(R.layout.fragment_main_me, container, false)
+    ) = inflater.inflate(R.layout.fragment_main_video, container, false)
 }
